+++ sched/semaphore/sem_trywait.c
@@ -64,7 +64,7 @@ static int nxsem_trywait_slow(FAR sem_t *sem)
 {
   FAR struct tcb_s *rtcb;
   irqstate_t flags;
-  short semcount;
+  int semcount;
   int ret;
 
   /* The following operations must be performed with interrupts disabled
@@ -153,7 +153,7 @@ int nxsem_trywait(FAR sem_t *sem)
 #if !defined(CONFIG_PRIORITY_INHERITANCE) && !defined(CONFIG_PRIORITY_PROTECT)
   if (sem->flags & SEM_TYPE_MUTEX)
     {
-      short old = 1;
+      int old = 1;
       if (atomic_compare_exchange_weak_explicit(NXSEM_COUNT(sem), &old, 0,
                                                 memory_order_acquire,
                                                 memory_order_relaxed))
